project: move SDL_mixer.h into SDL3_mixer subfolder
