+++ VisualC/SDL_mixer.vcxproj.filters
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClInclude Include="..\include\SDL_mixer.h">
+    <ClInclude Include="..\include\SDL3_mixer\SDL_mixer.h">
       <Filter>Public Headers</Filter>
     </ClInclude>
     <ClInclude Include="..\src\utils.h">
