+++ build-scripts/android-prefab.sh
@@ -42,9 +42,9 @@ android_api=19
 android_ndk=21
 android_stl="c++_shared"
 
-sdlmixer_major=$(sed -ne 's/^#define SDL_MIXER_MAJOR_VERSION  *//p' "${sdlmixer_root}/include/SDL3/SDL_mixer.h")
-sdlmixer_minor=$(sed -ne 's/^#define SDL_MIXER_MINOR_VERSION  *//p' "${sdlmixer_root}/include/SDL3/SDL_mixer.h")
-sdlmixer_patch=$(sed -ne 's/^#define SDL_MIXER_PATCHLEVEL  *//p' "${sdlmixer_root}/include/SDL3/SDL_mixer.h")
+sdlmixer_major=$(sed -ne 's/^#define SDL_MIXER_MAJOR_VERSION  *//p' "${sdlmixer_root}/include/SDL3_mixer/SDL_mixer.h")
+sdlmixer_minor=$(sed -ne 's/^#define SDL_MIXER_MINOR_VERSION  *//p' "${sdlmixer_root}/include/SDL3_mixer/SDL_mixer.h")
+sdlmixer_patch=$(sed -ne 's/^#define SDL_MIXER_PATCHLEVEL  *//p' "${sdlmixer_root}/include/SDL3_mixer/SDL_mixer.h")
 sdlmixer_version="${sdlmixer_major}.${sdlmixer_minor}.${sdlmixer_patch}"
 echo "Building Android prefab package for SDL_mixer version $sdlmixer_version"
