+++ build-scripts/test-versioning.sh
@@ -9,7 +9,7 @@ cd `dirname $0`/..
 # Needed so sed doesn't report illegal byte sequences on macOS
 export LC_CTYPE=C
 
-header=include/SDL3/SDL_mixer.h
+header=include/SDL3_mixer/SDL_mixer.h
 ref_major=$(sed -ne 's/^#define SDL_MIXER_MAJOR_VERSION  *//p' $header)
 ref_minor=$(sed -ne 's/^#define SDL_MIXER_MINOR_VERSION  *//p' $header)
 ref_micro=$(sed -ne 's/^#define SDL_MIXER_PATCHLEVEL  *//p' $header)
