+++ examples/playwave.c
@@ -29,7 +29,7 @@
 
 #include <SDL3/SDL.h>
 #include <SDL3/SDL_main.h>
-#include <SDL3/SDL_mixer.h>
+#include <SDL3_mixer/SDL_mixer.h>
 
 #ifdef HAVE_SIGNAL_H
 #include <signal.h>
