+++ src/effect_position.c
@@ -25,7 +25,7 @@
 */
 
 #include <SDL3/SDL_endian.h>
-#include <SDL3/SDL_mixer.h>
+#include <SDL3_mixer/SDL_mixer.h>
 
 #include "mixer.h"
