+++ src/effect_stereoreverse.c
@@ -24,7 +24,7 @@
   effect callback API. They are meant for speed over quality.  :)
 */
 
-#include <SDL3/SDL_mixer.h>
+#include <SDL3_mixer/SDL_mixer.h>
 
 #define MIX_INTERNAL_EFFECT__
 #include "effects_internal.h"
