+++ src/mixer.c
@@ -21,7 +21,7 @@
 
 #include <SDL3/SDL.h>
 
-#include <SDL3/SDL_mixer.h>
+#include <SDL3_mixer/SDL_mixer.h>
 #include "mixer.h"
 #include "music.h"
 #include "load_aiff.h"
